WEB-2397 Changed the type of the value number field criteria result for better precision

+++ fid-core/src/main/java/com/n4systems/api/conversion/event/EventToModelConverter.java
@@ -215,7 +215,7 @@ public class EventToModelConverter implements ViewToModelConverter<Event, EventV
 
 			@Override
 			public CriteriaResult populate(NumberFieldCriteriaResult result) {
-	        	result.setValue(Float.parseFloat(criteriaResultView.getResultString()));
+	        	result.setValue(Double.parseDouble(criteriaResultView.getResultString()));
 				return result;
 			}
